@@ -2,7 +2,7 @@
   "name": "typescript",
   "author": "Microsoft Corp.",
   "homepage": "https://www.typescriptlang.org/",
-  "version": "3.9.10",
+  "version": "4.4.2",
   "license": "Apache-2.0",
   "description": "TypeScript is a language for application scale JavaScript development",
   "keywords": [
@@ -52,26 +52,26 @@
     "@types/q": "latest",
     "@types/source-map-support": "latest",
     "@types/through2": "latest",
-    "@types/travis-fold": "latest",
     "@types/xml2js": "^0.4.0",
-    "@typescript-eslint/eslint-plugin": "2.18.0",
-    "@typescript-eslint/experimental-utils": "2.18.0",
-    "@typescript-eslint/parser": "2.18.0",
+    "@typescript-eslint/eslint-plugin": "^4.28.0",
+    "@typescript-eslint/experimental-utils": "^4.28.0",
+    "@typescript-eslint/parser": "^4.28.0",
     "async": "latest",
-    "azure-devops-node-api": "^8.0.0",
+    "azure-devops-node-api": "^10.1.0",
     "browser-resolve": "^1.11.2",
     "browserify": "latest",
     "chai": "latest",
     "chalk": "latest",
     "convert-source-map": "latest",
     "del": "5.1.0",
-    "eslint": "6.8.0",
-    "eslint-formatter-autolinkable-stylish": "1.1.1",
-    "eslint-plugin-import": "2.20.0",
-    "eslint-plugin-jsdoc": "21.0.0",
+    "diff": "^4.0.2",
+    "eslint": "7.12.1",
+    "eslint-formatter-autolinkable-stylish": "1.1.4",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
     "eslint-plugin-no-null": "1.0.2",
     "fancy-log": "latest",
-    "fs-extra": "^6.0.1",
+    "fs-extra": "^9.0.0",
     "glob": "latest",
     "gulp": "^4.0.0",
     "gulp-concat": "latest",
@@ -87,7 +87,6 @@
     "mocha-fivemat-progress-reporter": "latest",
     "ms": "latest",
     "node-fetch": "^2.6.0",
-    "playwright": "0.12.1",
     "plugin-error": "latest",
     "pretty-hrtime": "^1.0.3",
     "prex": "^0.4.3",
@@ -95,8 +94,7 @@
     "remove-internal": "^2.9.2",
     "source-map-support": "latest",
     "through2": "latest",
-    "travis-fold": "latest",
-    "typescript": "^3.9.3",
+    "typescript": "^4.2.3",
     "vinyl": "latest",
     "vinyl-sourcemaps-apply": "latest",
     "xml2js": "^0.4.19"
@@ -129,5 +127,7 @@
     "source-map-support": false,
     "inspector": false
   },
-  "dependencies": {}
+  "volta": {
+    "node": "14.15.5"
+  }
 }
