@@ -1,63 +1,58 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const fs_1 = __importDefault(require("fs"));
-const path_1 = __importDefault(require("path"));
-const exec_1 = require("@actions/exec");
-const io_1 = require("@actions/io");
-const archiver_1 = __importDefault(require("archiver"));
-const ava_1 = __importDefault(require("ava"));
-const codeql_1 = require("./codeql");
-ava_1.default("unbundle creates a stable directory name", async (t) => {
-    const tmpDir = fs_1.default.mkdtempSync("tmp");
+import fs from "fs";
+import path from "path";
+import { exec } from "@actions/exec";
+import { rmRF } from "@actions/io";
+import archiver from "archiver";
+import test from "ava";
+import { runQuery, unbundleDatabase } from "./codeql";
+test("unbundle creates a stable directory name", async (t) => {
+    const tmpDir = fs.mkdtempSync("tmp");
     const cwd = process.cwd();
     process.chdir(tmpDir);
     try {
         const testText = "hello world";
         const dbFile = "example.zip";
-        const output = fs_1.default.createWriteStream(dbFile);
-        const archive = archiver_1.default("zip");
+        const output = fs.createWriteStream(dbFile);
+        const archive = archiver("zip");
         archive.pipe(output);
         archive.append(testText, { name: "original-database-name/file.txt" });
         await archive.finalize();
-        await codeql_1.unbundleDatabase(dbFile);
-        t.is(fs_1.default.readFileSync(path_1.default.join("database", "file.txt"), "utf-8"), testText);
+        await unbundleDatabase(dbFile);
+        t.is(fs.readFileSync(path.join("database", "file.txt"), "utf-8"), testText);
     }
     finally {
         process.chdir(cwd);
-        await io_1.rmRF(tmpDir);
+        await rmRF(tmpDir);
     }
 });
-ava_1.default("running a basic query", async (t) => {
-    const tmpDir = fs_1.default.mkdtempSync("tmp");
+test("running a basic query", async (t) => {
+    const tmpDir = fs.mkdtempSync("tmp");
     const cwd = process.cwd();
     process.chdir(tmpDir);
     try {
         const projectDir = "project";
         const dbDir = "db";
-        fs_1.default.mkdirSync(projectDir);
-        const testFile = path_1.default.join(projectDir, "test.js");
-        fs_1.default.writeFileSync(testFile, "const x = 1;");
-        await exec_1.exec("codeql", [
+        fs.mkdirSync(projectDir);
+        const testFile = path.join(projectDir, "test.js");
+        fs.writeFileSync(testFile, "const x = 1;");
+        await exec("codeql", [
             "database",
             "create",
             "--language=javascript",
             `--source-root=${projectDir}`,
             dbDir,
         ]);
-        await codeql_1.runQuery("codeql", "javascript", dbDir, "import javascript\nfrom File f select f", "a/b");
-        t.true(fs_1.default
-            .readFileSync(path_1.default.join("results", "results.md"), "utf-8")
+        await runQuery("codeql", "javascript", dbDir, "import javascript\nfrom File f select f", "a/b");
+        t.true(fs
+            .readFileSync(path.join("results", "results.md"), "utf-8")
             .includes(testFile));
-        t.true(fs_1.default.existsSync(path_1.default.join("results", "results.bqrs")));
-        t.true(fs_1.default.existsSync(path_1.default.join("results", "results.csv")));
-        t.is(fs_1.default.readFileSync(path_1.default.join("results", "nwo.txt"), "utf-8"), "a/b");
+        t.true(fs.existsSync(path.join("results", "results.bqrs")));
+        t.true(fs.existsSync(path.join("results", "results.csv")));
+        t.is(fs.readFileSync(path.join("results", "nwo.txt"), "utf-8"), "a/b");
     }
     finally {
         process.chdir(cwd);
-        await io_1.rmRF(tmpDir);
+        await rmRF(tmpDir);
     }
 });
 //# sourceMappingURL=codeql.test.js.map
