@@ -1,12 +1,7 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.interpret = exports.toMd = exports.toS = void 0;
-const events_1 = require("events");
-const stream_1 = __importDefault(require("stream"));
-const util_1 = require("util");
+import { once } from "events";
+import stream from "stream";
+import { promisify } from "util";
+export { toS, toMd, interpret };
 // If e is an object, then we assume it is a single entity result of the form
 // produced by `codeql bqrs decode --format=json`. For example:
 //
@@ -43,15 +38,13 @@ function toS(e, nwo, src, ref) {
     url = `[${e.label}](${url})`;
     return url;
 }
-exports.toS = toS;
 function toMd(tuple, nwo, src, ref) {
     return `| ${tuple.map((e) => toS(e, nwo, src, ref)).join(" | ")} |\n`;
 }
-exports.toMd = toMd;
-const finished = util_1.promisify(stream_1.default.finished);
+const finished = promisify(stream.finished);
 async function write(output, s) {
     if (!output.write(s)) {
-        await events_1.once(output, "drain");
+        await once(output, "drain");
     }
 }
 async function interpret(output, results, nwo, src, ref) {
@@ -67,5 +60,4 @@ async function interpret(output, results, nwo, src, ref) {
     output.end();
     return finished(output);
 }
-exports.interpret = interpret;
 //# sourceMappingURL=interpret.js.map
