@@ -1,4 +1,3 @@
-"use strict";
 /*
 **************************************************
 ** CODE GENERATED BY: https://app.quicktype.io/ **
@@ -19,11 +18,9 @@ visit: https://app.quicktype.io/
 - Copy code
 - Paste below (and typescript compiler will perhaps make you remove some unused so check)
 */
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Convert = void 0;
 // Converts JSON strings to/from your types
 // and asserts the results of JSON.parse at runtime
-class Convert {
+export class Convert {
     static toJSONResult(json) {
         return cast(JSON.parse(json), r("JSONResult"));
     }
@@ -31,7 +28,6 @@ class Convert {
         return JSON.stringify(uncast(value, r("JSONResult")), null, 2);
     }
 }
-exports.Convert = Convert;
 function invalidValue(typ, val, key = '') {
     if (key) {
         throw Error(`Invalid value for key "${key}". Expected type ${JSON.stringify(typ)} but got ${JSON.stringify(val)}`);
