@@ -6,7 +6,7 @@ import { rmRF } from "@actions/io";
 import archiver from "archiver";
 import test from "ava";
 
-import { runQuery, unbundleDatabase } from "./codeql";
+import { runQuery, unbundleDatabase } from "./codeql.js";
 
 test("unbundle creates a stable directory name", async (t) => {
   const tmpDir = fs.mkdtempSync("tmp");
