@@ -2,8 +2,8 @@ import { Stream } from "stream";
 
 import test from "ava";
 
-import { toS, toMd, interpret } from "./interpret";
-import { Convert } from "./json-result-generated";
+import { toS, toMd, interpret } from "./interpret.js";
+import { Convert } from "./json-result-generated.js";
 
 const badResults = `{
   "#select": {}
